+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix fair share worker choice strategy internals update: ensure virtual task end timestamp is computed at task submission.
+
 ## [2.4.12] - 2023-05-06
 
 ### Added